♻️ Refactor(character): Standardize player and character dto form schemas
