+++ backend/src/resources/character/player/dto/appearance/appearance.dto.ts
