+++ backend/src/resources/character/player/dto/class/class.dto.ts
