+++ backend/src/resources/character/player/dto/progression/progression.dto.ts
