+++ backend/src/resources/character/schemas/stats/sub/abilityScores.schema.ts
