+++ backend/src/resources/character/schemas/stats/sub/skill.schema.ts
