+++ backend/src/resources/character/schemas/stats/sub/speed.schema.ts
